[Bronze IV] Title: 숫자의 합, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/11720. 숫자의 합/숫자의 합.cc
@@ -9,7 +9,7 @@ int main()
     cin >> n >> s;
     for(int i=0; i<n; i++)
     {
-        sum += ((int)s[i]-48);
+        sum += (s[i]-48);
     }
     cout << sum;
     return 0;
